woodlens geri tepme eklenecek

+++ Assets/Scripts/Manager/MiniBomp/MiniBompManager.cs
@@ -7,7 +7,7 @@ public class MiniBompManager : MonoBehaviour
     public static MiniBompManager miniBompManager;
 
     public int speed;
-    public int spawnSpeed;
+    public float spawnSpeed;
     public int range;
 
     private void Awake()
